+++ tests/testapp/admin.py
@@ -1,3 +1,5 @@
+from typing import ClassVar
+
 from django.contrib import admin
 
 from .models import Article
@@ -5,4 +7,9 @@ from .models import Article
 
 @admin.register(Article)
 class ArticleAdmin(admin.ModelAdmin):
-    pass
+
+    actions: ClassVar = ['rename']
+
+    @admin.action(description="rename it")
+    def rename(self, request, queryset):
+        queryset.update(title="renamedfine")