+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # pytest-djangoapp changelog
 
 
-### Unreleased
+### v1.5.0 [2025-09-27]
 * ++ Utils. Added 'admin_client' fixture.
 * ++ Utils. Added 'test_registered_urls' fixture.